SDET-HTML/requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.3.5
aiohttp==3.10.2
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
asttokens==2.4.1
attrs==24.2.0
azure-core==1.30.2
azure-identity==1.17.1
babel==2.16.0
backcall==0.2.0
bcrypt==4.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
build==1.2.1
CacheControl==0.14.0
cachetools==5.4.0
captcha==0.6.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
courlan
cryptography==43.0.0
cycler==0.12.1
dataclasses-json==0.6.7
dateparser==1.2.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
dirtyjson==1.0.8
distro==1.9.0
docopt==0.6.2
executing==2.0.1
extra-streamlit-components==0.1.71
fastapi==0.112.0
fastjsonschema==2.20.0
ffmpy==0.4.0
filelock==3.15.4
firebase-admin==6.5.0
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
gcloud==0.18.3
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.4
google-api-core==2.19.1
google-api-python-client==2.140.0
google-auth==2.33.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.17.0
google-cloud-storage==2.18.2
google-crc32c==1.5.0
google-generativeai==0.5.4
google-resumable-media==2.7.2
googleapis-common-protos==1.63.2
gradio
gradio_client==1.1.1
greenlet==3.0.3
grpcio==1.65.4
grpcio-status==1.62.3
h11==0.14.0
htmldate
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.24.5
idna==3.7
importlib_resources==6.4.0
ipython>=7.34.0,<8.12.3
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
jusText==3.0.1
jwcrypto==1.5.6
kiwisolver==1.4.5
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.52
langchain-text-splitters==0.0.2
langsmith==0.1.98
lavague==1.1.19
lavague-contexts-openai==0.2.4
lavague-core==0.2.33
lavague-drivers-selenium==0.2.12
lavague-gradio==0.2.8
llama-cloud==0.0.13
llama-index==0.10.56
llama-index-agent-openai==0.2.9
llama-index-cli==0.1.13
llama-index-core==0.10.56
llama-index-embeddings-azure-openai==0.1.11
llama-index-embeddings-gemini==0.1.8
llama-index-embeddings-openai==0.1.11
llama-index-indices-managed-llama-cloud==0.2.7
llama-index-legacy==0.9.48
llama-index-llms-azure-openai==0.1.10
llama-index-llms-gemini==0.1.11
llama-index-llms-openai==0.1.26
llama-index-multi-modal-llms-azure-openai==0.1.4
llama-index-multi-modal-llms-gemini==0.1.7
llama-index-multi-modal-llms-openai==0.1.9
llama-index-program-openai==0.1.6
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.33
llama-index-readers-llama-parse==0.1.6
llama-index-retrievers-bm25==0.1.5
llama-parse==0.4.9
lxml==5.2.2
lxml_html_clean==0.1.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib==3.9.1.post1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
msal==1.30.0
msal-extensions==1.2.0
msgpack==1.0.8
multidict==6.0.5
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.2
numpy==1.26.4
oauth2client==4.1.3
openai==1.40.2
orjson==3.10.7
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pickleshare==0.7.5
pillow==10.3.0
pip==24.0
pip-tools==7.4.1
# pipreqs==0.5.0
platformdirs==4.2.2
portalocker==2.10.1
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==4.25.4
pure_eval==0.2.3
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
pydub==0.25.1
Pygments==2.18.0
PyJWT==2.9.0
pyparsing==3.1.2
pypdf==4.3.1
pyproject_hooks==1.1.0
Pyrebase4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jwt==4.1.0
python-multipart==0.0.9
pytz==2024.1
# pywin==0.3.1
PyYAML==6.0.2
pyzmq==26.1.0
pexpect==4.9.0
ptyprocess==0.7.0
rank-bm25==0.2.2
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
requests-toolbelt==0.10.1
rich==13.7.1
rpds-py==0.20.0
rsa==4.9
ruff==0.5.7
seaborn==0.13.2
selenium==4.23.1
semantic-version==2.10.0
setuptools==72.1.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.32
stack-data==0.6.3
starlette==0.37.2
streamlit==1.37.1
streamlit-authenticator==0.3.3
streamlit-lottie==0.0.5
streamlit-option-menu==0.3.13
streamlit-react-flow==0.0.3
striprtf==0.0.26
tenacity==8.3.0
tiktoken==0.7.0
tinycss2==1.3.0
tld==0.13
toml==0.10.2
tomlkit==0.12.0
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.5
trafilatura==1.12.0
traitlets==5.14.3
trio==0.26.2
trio-websocket==0.11.1
typer==0.12.3
typing_extensions==4.12.2
typing-inspect==0.9.0
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
urllib3>=1.21.1,<3
uvicorn==0.30.5
watchdog==4.0.1
wcwidth==0.2.13
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
websockets==11.0.3
wheel==0.44.0
wrapt==1.16.0
wsproto==1.2.0
yarg==0.1.9
yarl==1.9.4
